Remove duplicate defines of StreamOrDevice and is_big_endian
